git-svn-id: https://peerblock.googlecode.com/svn/branches/0.9.x@83 d3e0e2f6-51e0-11de-b413-c5e99dc0a8e2

+++ pg2/updatelists.h
@@ -32,4 +32,4 @@ int UpdateLists(HWND parent);
 // Blocks without updating lists
 void BlockWithoutUpdate(HWND hwnd);
 
-extern const unsigned int g_build;
+extern const unsigned long long g_build;
